moving the index template
